Sort rooms

+++ Hotel1/src/com/senlainc/git_courses/java_training/petushok_valiantsin/service/RoomService.java
@@ -2,6 +2,7 @@ package com.senlainc.git_courses.java_training.petushok_valiantsin.service;
 
 import com.senlainc.git_courses.java_training.petushok_valiantsin.api.repository.IRoomDao;
 import com.senlainc.git_courses.java_training.petushok_valiantsin.api.service.IRoomService;
+import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Order;
 import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Room;
 import com.senlainc.git_courses.java_training.petushok_valiantsin.model.status.Free;
 import com.senlainc.git_courses.java_training.petushok_valiantsin.model.status.Rented;
@@ -14,9 +15,9 @@ import java.util.Comparator;
 
 public class RoomService implements IRoomService {
     private final IRoomDao roomDao;
-//    private final Comparator<Room> SORT_BY_PRICE = Comparator.comparing(firstRoom -> String.valueOf(firstRoom.getPrice()));
-//    private final Comparator<Room> SORT_BY_CLASSIFICATION = Comparator.comparing(Room::getClassification);
-//    private final Comparator<Room> SORT_BY_ROOM_NUMBER = Comparator.comparing(firstRoom -> String.valueOf(firstRoom.getRoomNumber()));
+    private final Comparator<Room> SORT_BY_PRICE = Comparator.comparing(firstRoom -> String.valueOf(firstRoom.getPrice()));
+    private final Comparator<Room> SORT_BY_CLASSIFICATION = Comparator.comparing(Room::getClassification);
+    private final Comparator<Room> SORT_BY_ROOM_NUMBER = Comparator.comparing(firstRoom -> String.valueOf(firstRoom.getRoomNumber()));
 
     public RoomService(IRoomDao roomDao) {
         this.roomDao = roomDao;
@@ -28,11 +29,11 @@ public class RoomService implements IRoomService {
 
     @Override
     public void delete(int index) {
-        try {
-            roomDao.delete(index);
-        } catch (ArrayIndexOutOfBoundsException e) {
+        if(roomDao.readAll().size() < index) {
             System.err.println("Room with index: " + index + " dont exists.");
+            return;
         }
+        roomDao.delete(index);
     }
 
     @Override
@@ -84,6 +85,13 @@ public class RoomService implements IRoomService {
         }
     }
 
+    @Override
+    public void show(MyList<Room> myList) {
+        for (int i = 0; i < myList.size(); i++) {
+            System.out.print(myList.get(i));
+        }
+    }
+
     private void showAllRoom() {
         for (int i = 1; i <= roomDao.readAll().size(); i++) {
             System.out.print(roomDao.read(i));
@@ -109,56 +117,41 @@ public class RoomService implements IRoomService {
         System.out.println("\nNumber of free room: " + counter);
     }
 
-//    @Override
-//    public void sort(String parameter) {
-//        switch (parameter) {
-//            case "price":
-//                sortByPrice();
-//                break;
-//            case "classification":
-//                sortByClassification();
-//                break;
-//            case "room number":
-//                sortByRoomNumber();
-//                break;
-//        }
-//    }
+    @Override
+    public MyList<Room> sort(String parameter) {
+        MyList<Room> myList = new MyList<>();
+        createBufList(myList);
+        switch (parameter) {
+            case "price":
+                sortByPrice(myList);
+                return myList;
+            case "classification":
+                sortByClassification(myList);
+                return myList;
+            case "room number":
+                sortByRoomNumber(myList);
+                return myList;
+        }
+        return null;
+    }
 
-//    private void sortByPrice() {
-//        System.out.print("\nSort by price:");
-//        MyList<Room> myList = new MyList<>();
-//        for (int i = 0; i < roomDao.readAll().size(); i++) {
-//            myList.add(roomDao.read(i));
-//        }
-//        myList.sort(SORT_BY_PRICE);
-//        for (int i = 0; i < myList.size(); i++) {
-//            System.out.print(myList.get(i));
-//        }
-//    }
-//
-//    private void sortByClassification() {
-//        System.out.print("\nSort by classification:");
-//        MyList<Room> myList = new MyList<>();
-//        for (int i = 0; i < roomDao.readAll().size(); i++) {
-//            myList.add(roomDao.read(i));
-//        }
-//        myList.sort(SORT_BY_CLASSIFICATION);
-//        for (int i = 0; i < myList.size(); i++) {
-//            System.out.print(myList.get(i));
-//        }
-//    }
-//
-//    private void sortByRoomNumber() {
-//        System.out.print("\nSort by room number:");
-//        MyList<Room> myList = new MyList<>();
-//        for (int i = 0; i < roomDao.readAll().size(); i++) {
-//            myList.add(roomDao.read(i));
-//        }
-//        myList.sort(SORT_BY_ROOM_NUMBER);
-//        for (int i = 0; i < myList.size(); i++) {
-//            System.out.print(myList.get(i));
-//        }
-//    }
+    private void createBufList(MyList<Room> myList) {
+        for (int i = 1; i <= roomDao.readAll().size(); i++) {
+            myList.add(roomDao.read(i));
+        }
+    }
+
+    private void sortByPrice(MyList<Room> myList) {
+        myList.sort(SORT_BY_PRICE);
+    }
+
+    private void sortByClassification(MyList<Room> myList) {
+        myList.sort(SORT_BY_CLASSIFICATION);
+    }
+
+    private void sortByRoomNumber(MyList<Room> myList) {
+        myList.sort(SORT_BY_ROOM_NUMBER);
+    }
 
     @Override
     public void showAfterDate(LocalDate freeDate) {
